Fix merged_into_type_id column type to match litter_object_types.id
litter_object_types.id is bigint unsigned ($table->id()); the FK column was
declared unsignedInteger, so MySQL rejected the foreign key and the migrate
step failed on CI before any test ran. Caught by CI on PR #717.

+++ database/migrations/2026_09_08_000001_add_merged_into_clo_and_type_to_category_litter_object_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
     {
         Schema::table('category_litter_object', function (Blueprint $table) {
             $table->unsignedBigInteger('merged_into_clo_id')->nullable()->after('litter_object_id');
-            $table->unsignedInteger('merged_into_type_id')->nullable()->after('merged_into_clo_id');
+            $table->unsignedBigInteger('merged_into_type_id')->nullable()->after('merged_into_clo_id');
 
             $table->foreign('merged_into_clo_id')
                 ->references('id')->on('category_litter_object')
